Handle video session modify responses while paused.
There is a limitation in the modem implementation whereby when the remote
party has paused video (eg sent dialer to the background) any session
modify requests which are sent are responded to, but the video state of
the connection is NOT updated until the video state becomes unpaused
again.

To handle this, ImsPhoneConnection will now listen for session modify
responses received via the VideoProvider.  When paused, valid
responses to requests will be used to update the current video state of
the ImsPhoneConnection.  When the video becomes unpaused, we will resume
using the modem reported video state.

Bug: 29381561
Change-Id: I1ac854dbc292affb3bc0bb9ce98b2fea33014cbb