Stop OUT_OF_SERVICE message during Emergency Call
Currently, the radio can report an OUT_OF_SERVICE (but Emergency calls
are available) status from the radio during an emergency call if that
call is made when there is no SIM card available or the SIM card is
locked. If the remote party disconnects, we mistakenly show a "Cellular
Network not Available" dialog when a call ends successfully because the
disconnect cause is mapped to OUT_OF_SERVICE. We now check to see if the
call was an emergency call before mapping to that disconnect cause.

Bug: 30657253
Change-Id: I7bf16c502c12125506ed31a56278f68ea39737ac