Add support for downgrade to audio when data limit is reached.
Fix the SUPPORTS_DOWNGRADE_TO_VOICE calculation in ImsPhoneConnection; it
was relying on signaling from the modem, which apparently doesnt work.
Instead relying on a carrier config option to set that value.

Adding logic in ImsPhoneCallTracker to auto-downgrade to audio and send
a connection event indicating why it happeend to Dialer.  This is only used
when the carrier supports it.

Bug: 30702393
Change-Id: Ie7ed3ba8b8aa6c45391d9dd45c4bbf7fe8c957ee
(cherry picked from commit 2e7e09e79c04f34034bebea8f5d86606bb584577)