Add carrier regex filtering to VisualVoicemailSmsFilter
This CL adds a extra step in the VisualVoicemailSmsFilter which will
match the message body against the carrier specified regex. Even if
the message have a unknown format if it matches the regex it will
still be considered a VVM SMS.

For example, Verizon requires any SMS that starts with "//VZW" to be
dropped. Other cases are handled by the AppDirectedSms app, but
"//VZWVVM"  is left for the system visual voicemail. Previously
"//VZWVVM garbage data" will pass though the filter because it is not
a valid VVM SMS, which is against the specification of dropping
everything. After this CL it will match the regex and dropped.
The VVM client will also receive the message body, but ignore it.

Change-Id: I8d446d69d75914bf5fe63834cfebec92f7917711
Fixes: 30954955