Update patch set 10
Patch Set 10:

Ethan wanted it how this was originally, with the decrypt check, to stay in line with AOSP. They're handling the modem state differently whether it's encrypted or not.

You can check RIL version in radio log. If you grab it after a boot, it should be reported at the beginning.

Patch-set: 10