Update patch set 1
Patch Set 1:

Pretty sure it won't, as we account for that in the follow-up patches. There was a really lengthy conversation around this when we did these changes in CM13. Look up the original patches here on gerrit. Maybe Ethan can contribute with some more insight, as he was the one who helped getting it where it is.

Patch-set: 1