Update patch set 12
Patch Set 12:

Did not want to get to much involved in the legacy_ril stuff since I mainly cherrypicked the work from others to get radio on my moto g2.

I agree with arne on not holding this one for to long though. As expected with the patch series I got calls/data and both sim working on my phone.

Patch-set: 12