Fix ServiceStateTrackerTest and CdmaSmsCbTest
Use an unsupported encoding scheme in testCmasUnsupportedCharSet of
CdmaSmsCbTest. GCM_DCS is supported. So change it to 0x1F. Using 0x1F
since it encoding scheme is a 5 bit value and 0x1F is maximum.

Increase the wait time in ServiceStateTrackerTest to 200ms. Tests were a
bit flaky when the time was 100ms.

Change-Id: Ieac80b5b1b80819acd6fb314f48a5fada68db791