Update reference to config_restart_radio
For the switch to CarrierConfig, we update the method used to get the
value of what used to be called
config_restart_radio_on_pdp_fail_regular_deactivation to the new method
provided by CarrierConfig. Since we need to pass the context, we also
update the places this value is accessed so they pass their context.

Merged-In: Id43344339b07628d62e640bf73f75af048a1259b

Bug: 29873049
Change-Id: Iba091f43e094b823b5f27db6b5f05b5a362c4658