Fix answering waiting call simultaneously with active call disconnect
When the user answers a call-waiting call at the same time as the remote
end disconnects the formerly-active call, an error may occur where the
in-call ui is left in an inconsistent state and the waiting call is
never answered.

Normally, when a call-waiting call is answered, the active call will
first be held. Then, after the modem reports that as successful,
onCallHeld will issue the EVENT_RESUME_BACKGROUND message to answer the
call-waiting call. However, if the remote end disconnects while the user
is answering, onCallTerminated will be called instead.

This change alters onCallTerminated to issue the EVENT_RESUME_BACKGROUND
message if it occurs in the middle of a call-switch.

Change-Id: Icb757274ad63af46ebff5028dc484f8497777987
Fix: 31684640
(cherry picked from commit 3379e399780e6e6531eb84f14d1eca92ae48af2a)