Plug hole in Roaming indication.
If we change phone types after falling into a roaming condition
(due to other bug/race) we were tossing the old state and not
notifying listeners.  This could leave listeners thinking we
were still roaming.

Also added logging to try to catch this in the future and made
DataConnection (a listener) more robust by rechecking the roaming
state when it connects.

bug:31278802
Change-Id: Ica5b2bbf1cb6fde5bc11994d1a699c53ac562ade