Allow to define a default app to handle Wap Push
Android handles Wap Push messages by sending them to the default MMS
package, if one exists, and broadcasts it if not.

This patch creates an overlayable string which, if set, will announce
what should be the Wap Push receiver app. If that string (and package)
is available, then the message will be sent to it instead. If not, the
previous behavior is maintained: message will go to the default MMS
package (if there's one), or broadcasted.

Issue: PAELLA-286
Change-Id: I0e4624dab4e51abfc4e3e048c2d88932aa4af616