Load CarrierConfig and set SIM state to LOADED for NV subscription
With commit e8ed3af7e0dc ("CdmaServiceStateTracker: Load CarrierConfig
for NV Subscription"), CarrierConfig was loaded as soon as NV_READY
was received. With this patch CarrierConfig is loaded after the ICC
subscription has been updated as it is normally done. Moreover, the
use of updateCarrierServices() should fix CarrierService.

In addition to that, this patch causes the SIM state to transition
from READY to LOADED once the the ICC subscription has been updated.

Change-Id: I3332310d8493cc578b8f565e59c9805c8e2027a3