Only get WiFi status based on RAT extras
We used to set the WiFi Connection property based on the the provisioning
status of the phone until the RAT extra was set from the modem. This is an
obsolete method of determining whether or not to set the WiFi property and
was causing some calls to be improperly marked as WiFi.

Test: Manual Testing
Bug: 30964593
Change-Id: I9922ff17ffb704a09e9b2cebeb147a7463b9898b