Allow fallback translated languages in MccTable
When choosing a fallback locale, if that locale is marked as
"translated" by LocaleStore, return that locale. This is true if a
locale itself has translations, or another locale with the same
language and script that is translated.

For example if the user inserts a Canadian SIM, previously it will
fallback to en-US since the system doesn't have translations for
en-CA. But since Android N users are allowed to select locale
English (Canada), which the framework will find suitable fallback
translation when resolving resources. So with this change, if the
user inserts a Canadian SIM, getLocaleFromMcc will now return en-CA.

Test: Run MccTableTest (via adb shell am instrument)
Bug: 31395257
Change-Id: I7f74bc2ed14efb9ce5594dbf563c76e2d5b353da