SubscriptionInfoUpdater: Unbreak usage of TelephonyPlugin
Change I5b829662b466f609cd22adfa7d3b754757258463 made some changes
to how and when we accept to update the subInfo records which depend
on state flags and messages added the SubscriptionInfoUpdater.

However, a delegated TelephonyPlugin (say... QTI extensions) isn't
aware of those, won't manipulate them, and won't generate the expected
update messages, resulting in missed data updates and inconsistent
subscriptionInfo objects (in particular, as a result of hot swaps)

So... return the behavior of the exported updateSubscriptionInfoByIccId()
to the original expectations, and reserve the new behavior for internal
calls.

RM-290

Change-Id: I4c9e9c40756d049d346145cb03e0c3c5c84e13ad