Fixed that MMS incorrectly treated as metered data traffic.
The metered APN types cache was built before carrier config
is ready. Get rid of the cache because it doesn't buy us too much
in terms of performance as carrier config already has its own
cache. This will ensure we load the up to date configuration
when we bring up the data call.

Test: Manually sent mms on Vzw network.
bug: 31233558
Change-Id: Id0c243a7a92bd6ab3e7a46053e173bf727041015