DO NOT MERGE If set data profile with roaming protocol, set back to home when roaming off.
Data profile will be updated in O to support both home and roaming protocol.
This is a MR2 only fix to minimize behavior change and fix data error when
device move from roaming to home without airplane mode or power cycle.

The sequence of events can be tricky, eg. set data profile may be called before
device registers on network and will set home protocol even device is indeed roaming.
It is intentional this behavior is unchanged. Same for not set roaming protocol in
onRoamingOn.

Test: manual, https://paste.googleplex.com/4566715396521984
Bug: b/32240817
Change-Id: I70598bb14c7cb499552b1b725cf5b464f8ec3b2c