Fix In Conference MMI codes
According to the 3gpp spec, when in a GSM conference, one of the
participants can be disconnected when a specific MMI code is placed.
This code includes the index of the participant being disconnected.
When the participant is disconnected using this method, the GsmCdmaConnection
stays in the DISCONNECTED state in the GsmCdmaCall until all
GsmCdmaConnections in the conference are disconnected. This causes an
issue in GsmCdmaCallTracker#hangupConnectionByIndex because the
getGsmCdmaIndex() method will always return an Exception if the
connection is in the DISCONNECTED state.

We now explicitly check if the GsmCdmaConnection has been disconnected
before querying the connection index.

Test: Manual
Bug: 29588832
Change-Id: I5f71937fbaa75bb93aa2175791a01fa063cb12ac