Use video pause signaling when data limit reached.
When the data limit is reached and video is disabled, use the "pause"
video signaling instead of disconnecting the call.
Previously, ImsManager#updateImsServiceConfig would be called when data
was disabled due to hitting the data limit.  This would cause VT to be
unprovisioned; as a result it was not possible pause or unpause the video.
To solve this, when the data limit is reached and we're about to pause,
we are now deferring the updateImsServiceConfig until the call terminates.

Test: manual
Bug: 30760683
Change-Id: Ia1d4247cb5a11da7694e9d197a506442c7f675f4