Send Filtered VVM SMS to telephony with phone account handle
This follows the public VisualVoicemailService using the
PhoneAccountHandle as account key instead of subId, and uses
VisualVoicmailSms to wrap the SMS content.

Also, previous design of the filter sent the broadcast directly to the
VVM client, which was always telephony. After this CL it will always
send to telephony which will relay it to the client.

Test: CTS tests in follow up CL
Change-Id: I45a3d276af9cb6f9509586f05fc8fc600e074c6b