Update patch set 2
Patch Set 2:

I *believe* that this is required for kiwi to switch sims for mobile data.  I didn't test with the other 2 commits and this reverted to make sure it was really necessary but it seems like I should need this.

I'm happy to revert this and retest but I didn't want to do so pre-emptively...

Patch-set: 2
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0