Linger the carrier service binding when privileges drop.
If, when rebinding, no app has carrier privileges, but a binding is
active, instead of immediately unbinding, we schedule an unbind in 30
seconds. During that time period, if any other app comes up with
carrier privileges, the scheduled task is cancelled and the binding is
immediately dropped. If the same app and component regains carrier
privileges, the scheduled task is canceled and the binding left alive.

This keeps the binding alive through temporary blips in carrier
privileges. Note that apart from foreground status, the binding itself
profers no additional privileges or abilities onto the bound
application.

Bug: 33057511
Test: Remove/reinsert SIM before/after 30 seconds, disabling carrier
app entirely.
Change-Id: Ice194fe6ce58f1372369b1799d4d20bd5304f684