Properly route the SET_CALL_FORWARD_DONE message
In ImsPhone, the message indicating that call forwarding was set was
being directed straight to ImsPhoneMmiCode instead of routing through
the handler in ImsPhone, which caused the call forwarding notification
to never get set. This sends the proper message to ImsPhone first, which
will put up the call forwarding notification.

Change-Id: I5d06a0bad6e2bf368817ec623905f4a16e06d358
Fix: 34185152
Test: manual (to test on US carriers, force useImsForUt in GsmCdmaPhone)