Introduces ImsResolver and ImsServiceController Classes
First part in a series of changes to refactor IMS APIs and allow
multiple ImsServices to be managed based on device and carrier
configurations. This change includes:
1) Introduces the ImsResolver class, which caches the apps
defining an ImsService in their manifest. It then updates
which ImsServices it is bound to by monitoring package changes
and changes to the Carrier Configuration.
2) Introduces the ImsServiceController class, which manages
the binding to an ImsService as well as control changes to
the features that the ImsService supports.

Test: Adds unit tests
Bug: 30290416
Change-Id: Id5bb341903bcf03cf340a8552ad3a75e8409c494