Squashed merge of master-sim into master.
Includes the following commits:

==

Add stub implementation of EuiccController.

This class will provide the backing implementation for EuiccManager.

The service is only initialized if the eUICC hardware feature is
defined. This provides a simple gate to turn on/off the feature during
development.

==

Initial implementation of EuiccConnector.

EuiccConnector is a state machine responsible for finding and
connecting to, and then dispatching commands to the system
EuiccService implementation.

==

Update references to UiccAccessRule in UiccCarrierPrivilegeRules.

AccessRule is being moved to frameworks/base/telephony so that it can
be used to transfer access rules between an EuiccService
implementation and the platform.

==

Implementation of getEid() command.

Adds unit tests of command dispatch and bind lingering.

Also fixes a bug in ContextFixture that prevented bindService from
working when the Intent specified a component name but not a package
name.

==

Generalize EuiccConnector's resolution to work for activities.

Overall, resolving the active EuiccService implementation and
resolving the active Activity for a particular LUI intent action are
quite similar - both perform the same permission check on the app and
the activity/service declaration and require a priority. So we can
generalize the existing service lookup method to work for LUI
activities as well, and make this a utility method accessible to other
callers.

Also tweak the priority definition to use the more standard
android:priority instead of our own custom metadata. This reduces the
API surface and also ensures some sane behavior (like unbundled APK
updates not being able to adjust priorities without the OEM updating
the copy in /system in an OTA).

==

Euicc implementations for getting metadata and downloading.

These are fully implemented except for the following, which will be
handled as separate follow-up pieces:

-Carrier privilege checks (the privileged permission is currently a
strict requirement for the download API; this will be relaxed later
per the method spec)
-Resolution of user-resolvable errors

==

Test: TreeHugger
Change-Id: Ide3d81fdbdccff4573d58220fc2c603a7b9de5c4