Allow port and originating number to be specified for VVM SMS
This CL allows the VisualVoicemailService to ignore SMS not from the
designated number or to the designated port.

The VVM SMS filer can be configured to allow only SMS from certain
number or to certain port to be considered as a VVM SMS.
For numbers if the list is not empty, the originating number must be
an exact match to one of the numbers in the list.
For port it can be exact port, any data SMS port, or anything
(default).

Change-Id: I3d4aa34a342ba93d23841c07cbda2c2806f89dc4
Fixes: 36075510
Test: VisualVoicemailSmsFilterTest
      CTS VisualVoicemailServiceTest