Add extra data to auto network selection message.
If device is already in auto mode, and choose network automatically
option is selected, then network selection complete message is sent to the
application, which triggered the request, without sending the request
to RIL. On the app side, it is not known whether the request was
actually sent to RIL or not.

Add information to the message so that application knows if the request
has been sent to RIL or not.

Change-Id: Ieb5d87041aab2929c4478cb9139abc555114bbeb
Bug-Id: 37200906