Allow UTF-8 data SMS to be filtered as VVM SMS
The OMTP visual voicemail specification does not specify what the
VVM SMS is.

Previouly only data SMS that is encoded with the Data Coding Scheme
GSM 7bit or UCS2 can be recognized. These format cannot be sent with
SmsManager.sendDataMessage() and prevents CTS testing.

This CL will attempt to decode the 8bit binary DCS userdata as a UTF-8
string. This format can be sent with sendDataMessage()

Fixes: 33397325
Fixes: 36075510
Test: CTS VisualVoicemailServiceTest

Change-Id: I81aa2993d43d1a39e1192d02828a4d982475441f