Added ETWS primary message default message support
ETWS primary message does not contain messagy body. We used
to use hardcoded "ETWS" as the body, which is not easy for
end users to understand what happened. Added the built-in
default messages provided by Japanese government guideline
for earthquake, Tsunami, test, and other channels support to
enhance the user experience.

Test: manual
bug: 33595007
Change-Id: I0bd0a6655bc4a327124b2a145fa3892c177c4fdb