Call set initial attach apn when roaming status changes
This is needed for the old modem (RIL version <15) which only
accepts one protocol based on the roaming condition. In that
case we need to call set initial attach apn again so the
correct protocol can be passed down to the modem.

Test: Telephony sanity tests and unit tests
bug: 35210742
Merged-In: I5e88a61a1cc69d22ed319e598050509341833d7d
Change-Id: I5e88a61a1cc69d22ed319e598050509341833d7d