Relax VVM originating number matching
The visual voicemail client can restrict the originating number for
VVM SMS to reduce false positives. The matching criteria used to be
exact match, which is too strict. Especially on CTS tests where
the formatting of the number cannot be controlled.

In this CL PhoneNumberUtils.compare() is used instead, which will
ignore formatting and country codes.

Bug: 37106950
Test: CTS android.telephony.cts.VisualVoicemailServiceTest
Change-Id: I2d28a3718b1a87f8618d171d71a08809763b9715