Clean up Icc Refresh handling
1. When a SIM refresh is received of type INIT,
don't immediately start re-fetching the records, because
the app might not be ready. In this case, we could end up
broadcasting that all records are loaded pre-maturely.

On a refresh of the above type, the app will move out of the
ready state and back into the PIN state, if PIN is enabled.
Start reading the records only after the PIN has been verified
and the app has moved to the READY state.

2. When UiccController receives refresh of type INIT, it will
propagate it to the Card. The card will dispose the apps if required.
UiccController will also request card status on receiving
refresh to get the latest state.

This avoids having to make assumptions on the ordering of the
Unsols for Sim Refresh and card state changes. This also
ensures that after a refresh, the app state will move to ready
when it is truly ready.

3. Moved common implementation for refresh into IccRecords
for better code reuse.

Test: runtest -path frameworks/opt/telephony/tests/telephonytests/
Test: Manual SIM refresh related cases
Bug: 35919855
Change-Id: Ib4149c94adc7845bfe8d07dacec6ff78a6b64c1e