Implement remaining core eUICC APIs.
These APIs follow the template of the APIs which have already been
implemented, in general. Switch, delete, and update nickname take a
subscription ID and map this to an ICCID before passing it to the
EuiccService implementation.

This completes the baseline implementation as designed; any additional
APIs or tweaks to the APIs will be tracked as separate bugs.

Fixes: 33075886
Test: Build/boot/unit test
Change-Id: I98ad64fb70b8989d14f3f3b6f875e693258faf5d