Update patch set 1
Patch Set 1:

Was hoping to make this short typing from phone but oh well lol... On 7.1.1 I have a great LTE connection on 7.1.2 it takes the modem over a minute to finally come up and is stuck on 3g.  I saw the CDMA GSM flippy-floppy messages in logcats and thought this would be simple but the messages are gone after this patch but still stick...but it's looking like it's unrelated? Even tho for an incremental OS update that makes me cry a bit on the inside lol...

Patch-set: 1