Correct the value of access technology.
The byte in java is a signed variable.
The last byte in one record(bytes[offset + 4) will be 0xfxx if the first
byte in the next record is 0xf.
And the beginning in each record is 0xfffff to indicate the default PLMN.
To correct the value of accessTechs in the record, filtered the
bytes[offset + 4] with 0xff.

Fixes: 37884144
Change-Id: Ie88952321c4add09e6ad91f4a6db4e28d72daafc