IMS: Notify call failed for unobtainable number
"Server unreachable" is displayed when user makes IMS call with
unobtainable number. But "Invalid Number" is displayed when user makes
CS call with the same number. They are inconsistent. "Invalid Number"
should be displayed in IMS call.

To fix this, add a case that map the fail cause to
DisconnectCause#UNOBTAINABLE_NUMBER.

Test: manual - Checked that "Invalid number" is shown when make the IMS
MO call to unobtainable number.
Bug: 69940434

Change-Id: I7bf653c137ad0ee4aba0ab03ab4f054451410839