Fix EuiccControllerTest's dependency on activity manager internals.
verifyIntentSent() can't directly verify a call to
PendingIntent#send() using Mockito as PendingIntent is final. The
previous solution to this mocked out the highest level part of
PendingIntent's implementation which isn't final, but this part is an
implementation detail that is subject to breakage, so depending on it
is brittle. Instead, we can stub out the call to send() manually and
capture/verify the inputs in the test.

Test: bit FrameworksTelephonyTests:com.android.internal.telephony.euicc.EuiccControllerTest
Change-Id: I79a2213a4edf12b3359ba7425006694362a57854