Send user reject for SET_UP_CALL timeout
For the setup call command, a 'no response from user' is incorrectly
sent from the dialog activity when user doesn't respond to the dialog or
the dialog loses foreground focus.

Correct response per spec is 'user did not accept the proactive command'.
Route the error handling to the SET_UP_CALL handler to send this.

Bug: 38253898
Change-Id: I4f3263a2bf18979fb493da72039584f95c671d4d