Fixed unnecessary roaming off event triggered issue
ag/1489842 fixed an issue that when we are in roaming and
phone type suddenly switches, we didn't broadcast the
roaming off event and the previous roaming status is
tossed away. The fix ag/1489842 then always send the
roaming off status when phone type switches.

The new fix in this CL would be only broadcasting the
event when we need.

Test: Telephony sanity tests
bug: 37973830
Merged-In: I1cd1d027a7d56541422dc573ab2ecff1c6fde108
Change-Id: I1cd1d027a7d56541422dc573ab2ecff1c6fde108