IMS conference merge improvements.
1. Re-map ended by conference merge to "IMS_MERGED_SUCCESSFULLY".
This later gets translated to Telecom disconnect cause "OTHER".
This is helpful to ensure the InCall UI is able to recognize that a call
ended because of the fact it is joining a conference.
2. Send connection event when starting to merge two calls; this way the
InCall UI can suppress transient states when merging.
Note: I purposely only send "merge complete" when merge fails; that way
when the calls disconnect naturally at the end of the merge, the incall ui
still knows they're in a "merging" state.

Test: Manual
Bug: 37581781
Change-Id: I22abfbae470c7e67b6e6309f58c958820a697609