Cache the getActiveSubscriptionInfoList() in SubscriptionController
To fast reading the active subscription info list (the subinfo list
currently in the phone slotes), a cache is created and maintained inside
the SubscriptionController.

To protect cache, refresh the content of the cache every time there
is an update on the database, populate/regenerate the cache with
the updated information and thereby the reference of the cache object
will be changed.

When getting an Active Subscription Info list, get a deep copy of the
cache instead of its reference.

Test: Manual
Bug: 29877775
Change-Id: Ib26c550d63d4befa6f71ef82dfd6d73579b17e24