Fixed file descriptor leak in telephony unit tests
The handler thread in SimulatedCommands was not closed after
each unit test run. Fixed by adding a dispose method to quit
the thread.

Also use quit() instead of quitSafely() for all thread quiting
because we don't care if the messages were dispatched or not
after the test. This can fix some messages sending to dead
threads non-harmful exceptions in the output debug messages.

Test: Telephony unit tests
bug: 29579544
Change-Id: I112e46b0147a1950046a35cd9ac2498560c6ee1b