Notify carrier app of other default networks
When we need to disable data its common for the carrier to send the user
text messages explaining why with a URL.  The Carrier App will
url-clicks in this case and reroute them to a private network.  However,
it's possible for the user to get other sms's with links and if there is
another network to load them on we should not hijack them.  This CL
adds a new carrier action for carrier apps to register/unregister
notifications of other networks from ConnectivityService instead of
notifying only when carrier apps has disabled cell data. in that way
framework has decoupled the logic of network status registration from
cell data status.

Bug: 62487488
Test: Manual
Change-Id: I0a7b87ad07a0cf2fdb14f939e8f2865eedf7b008