Update subscription list when SIM state is NOT_READY.
This is required to support the eSIM boot profile, as NOT_READY is the
terminal state when this profile is active.

This also requires that we perform the update even if the IccId query
is not done, as this query don't be done in this case. The update
should nevertheless be safe since it's only doing database
manipulations local to the eSIM logic and not touching other state
within SubscriptionInfoUpdater.

Bug: 64113021
Test: TreeHugger, on-device switches and eSIM resets
Change-Id: I19c1e8d4b72aed28599178f73d335b3b10e61772