Update ImsPhone to have default OUT_OF_SERVICE state
ImsPhone only has the voice registration state POWER_OFF when
it is constructed. After it registers with IMS (or fails), it
moves to the STATE_IN_SERVICE and STATE_OUT_OF_SERVICE states.
There is no need to start it in the STATE_POWER_OFF voice
registration state and can instead be constructed in the
STATE_OUT_OF_SERVICE state.

Bug: 62114647
Test: Manual emergency call testing
Change-Id: I836548a460b6da70be9bf2807e4d791ee7724cc3