Clear calling identity in getCarrierImsPackage()
When the calling package has SEND_SMS but not READ_PHONE_STATE
the check currently errors out. This clears the calling package
so that we can still query the CarrierConfig to see if there is
an ImsPackage that should be checked for a CarrierSms intercept.
CarrierConfig expects READ_PHONE_STATE to check keys.

Bug: 37570992
Test: manual
Change-Id: I56a8336f9ab1fe0ad33e0e304431df125bbdeeec