Fix double prompt when downloading profiles.
Currently, if a carrier app wants to download a profile, two consent
dialogs may be triggered:

1. One to deactivate the current SIM and activate the eSIM, as we
cannot check for calling permissions without the eSIM being active.

2. Another to prompt the user for permission if the current carrier is
different, to prevent a carrier from hijacking cellular service.

Since the dialog for 2 is stronger than 1 (an assumption we were
already leveraging), when an unprivileged app requests a download,
skip straight to the second prompt even if the eSIM isn't active. If
the user accepts the prompt, we assume it's okay to switch to the eSIM
first and move forward.

In addition, we pass the calling app package name so the prompt UI can
show details of the app requesting consent.

Fixes: 38049463
Test: TreeHugger + E2E verification
Change-Id: Iaeb33f8a88aec92f323d91586e45d4852bb7a489