Handles the RIL error from the CommandException
In RadioResponse.java, all the RIL errors will be converted into
CommandException and sent back to the caller with the callback message.
So the AsyncResult in the callback message of 'EVENT_START_NETWORK_SCAN_DONE'
and 'EVENT_START_NETWORK_SCAN_DONE' will either contains a NetworkScanResult
without any exception, or contains an exception without any NetworkScanResult.
So if there is any error when trying to start/stop a scan, the error
code must be extracted from the CommandException.

Test: Telephony sanity tests
Change-Id: I9766b7fd530e9146e6d9867478ed79ad8f7ba1fc