Fix emergency call failure over IMS after boot
With some recent changes to the ImsPhone, we now never go into
the STATE_POWER_OFF state. This was how the GsmCdmaPhone used
to verify that the ImsPhone was not ready to make calls, so
it was not falling back to CS in this small space of time
before the ImsService came up and instead failing to make the call.

This change fixes this issue in two ways:
1) Instead of looking for the STATE_POWER_OFF state, it will instead
check to see if the ImsService is up.
2) If the emergency call fails over IMS, then fallback to CS
automatically.

Bug: 62618800
Test: Place emergency call right after boot up
Change-Id: Ia290681983e995568d01e84fa81d38d4b986d9b0
(cherry picked from commit 37414e6f5ddbbb8468607484d49169ae4d6b2801)