Suppress Connection address changes while merge is in progress.
Due to the manner in which conferences are formed, the IMS session
associated with the conference is swapped into the call which initiated
the merge briefly before the conference is added to Telecom and the
session is moved to that conference.  The issue is that in some cases
the new conference session will have a conference URI associated with it.
To mitigate this, this change blocks address changes during the merge
process.

Test: Manual
Bug: 62017197
Change-Id: I7e7c21e26380c55a6b4ed34ad50f384be7df9d5f