Suppress conference initiation when already in the process of merging.
When a conference merge is already in process, skip initiation of a
conference merge requested by the user.  This ensures if the user hammers
on the merge button in the UI or mistakenly double-taps, the modem won't
try to merge multiple times, one which will fail and generate an error
to the user.

Also cleaned up the log messages a wee bit to show the call IDs of the
calls merging.

Test: Manual
Bug: 38151476
Change-Id: I8a7e0d104b10d1d1cfbc6a76d8bf62a0adfa85c5