Proper handling of roaming override case
If the roaming state is overridden by the framework, we should
pass allowRoaming=true down to the modem so the data call will
not be rejected. Also we should pass the actual roaming state
to ril_service so it can pick the correct protocol (for old
modem backward compatibility support).

Test: Telephony sanity tests
bug: 62807105, 62285365
Change-Id: Ida93e6aa3949c8434aa58468f12e4512c9a02351