Fixed incorrect time zone when the device is on IWLAN
If the device is on IWLAN, modems manufacture a ServiceState
with the MCC/MNC of the SIM as if we were talking to towers.
Telephony code then uses that with mccTable to suggest a timezone.
We shouldn't do that if the MCC/MNC is from IWLAN.

Test: Telephony sanity tests
bug: 62563032
Merged-In: I7e5463909854c68ec57feae198071c73b0c5571c
Change-Id: I7e5463909854c68ec57feae198071c73b0c5571c