Do not set the state to POWER_OFF when in APM for ImsPhone
There is no need to set the ImsPhone to the state POWER_OFF
when we move into APM, since there is no difference between that
and the OUT_OF_SERVICE state.

Also, removes the POWER_OFF check from the ImsPhoneCallTracker,
as it should only determine whether or not it can dial a call
from the ImsService capabilities at the time of the call.

Bug:62816248
Test: Manual, Test E911 call
Change-Id: I904be6a8e79d40ebf8f8dae81277169ad7dabfcc