Add workaround for modem issue where pause bit is not set when on hold.
When a call is remotely held, the modem is not setting the pause bit on
the video state as it should.  This causes a cascading failure where the
user cannot turn the camera on/off (because another video pause hack is
not able to detect that the video is paused and thus activate the hack).

Test: Manual
Change-Id: I2a83b73df5488454b7030d7e4f7c78f360a27eda
Merged-In: Id30c5839b1e8c0d419ebcf03045ad1bcfe3b7d57
Fixes: 63153661