Ensure mShouldIgnoreVideoStateChanges is set on/off on fake pause/resume.
This is a fix for a modem bug where video state changes are not relayed
to the framework.
When I made the initial fix to add the changetoPausedState methods,
I thought I'd fixed the bug but it turns out I forget to set the
mShouldIgnoreVideoStateChanges variable.  There is a bit of a race
condition when reproducing this bug and I hit the race condition and
erroneously thought I'd solved it.

Test: Manual
Bug: 63153661
Change-Id: I82c9cb7bcf57df8f86964a3bff37469f0b999f1a