IMS: Reset Ims capabilities when feature state becomes not available.
If device is supporting only one IMS stack, when DDS is moved from
Sub1 to Sub2 (non-IMS), feature state of Sub1 becomes not available.

When feature state is not available, registration listener will be removed.

So if the IMS registration state change and service status updates are
received from modem after the listner is removed, feature capabilities
are not updated to phone account. As a result, make video call option
is shown on dialer UI, though there is no ims registration and video
capability for any of the subscriptions.

To fix this, reset the ims capabilities when ims feature state becomes
not available for that subscription.

Test: Manual
Bug: 64098630
Change-Id: I7b59a48539560e0bf40e41799a4b6513424db6c7