Remove sysprop based radio reset sync mechanism
This CL removes the gsm.radioreset system property that
could ostensibly be used to initiate an out-of-band
radio reset, which relies on both the framework and OEM
code to synchronously write the value of the gsm.radioreset
system property.

This property isn't being used by any vendor implementations
that are currently active, unless that usage is obfuscated
within binaries.

Bug: 64444973
Test: compilation
Change-Id: I5837d01d1d949ebf57a2d25f9a6fb52ed9f6db94